+++ ormar/models/model.py
@@ -5,6 +5,7 @@ from typing import (
     Set,
     TYPE_CHECKING,
     Tuple,
+    TypeVar,
     Union,
     TypeVar,
     Type,
@@ -19,6 +20,8 @@ from ormar.models.model_row import ModelRow
 if TYPE_CHECKING:  # pragma nocover
     from ormar import QuerySet
 
+T = TypeVar("T", bound="Model")
+
 
 class Model(ModelRow):
     __abstract__ = False
@@ -30,7 +33,7 @@ class Model(ModelRow):
         _repr = {k: getattr(self, k) for k, v in self.Meta.model_fields.items()}
         return f"{self.__class__.__name__}({str(_repr)})"
 
-    async def upsert(self, **kwargs: Any) -> "Model":
+    async def upsert(self: T, **kwargs: Any) -> T:
         """
         Performs either a save or an update depending on the presence of the pk.
         If the pk field is filled it's an update, otherwise the save is performed.
@@ -45,7 +48,7 @@ class Model(ModelRow):
             return await self.save()
         return await self.update(**kwargs)
 
-    async def save(self) -> "Model":
+    async def save(self: T) -> T:
         """
         Performs a save of given Model instance.
         If primary key is already saved, db backend will throw integrity error.
@@ -191,7 +194,7 @@ class Model(ModelRow):
             update_count += 1
         return update_count, visited
 
-    async def update(self, **kwargs: Any) -> "Model":
+    async def update(self: T, **kwargs: Any) -> T:
         """
         Performs update of Model instance in the database.
         Fields can be updated before or you can pass them as kwargs.
@@ -250,7 +253,7 @@ class Model(ModelRow):
         await self.signals.post_delete.send(sender=self.__class__, instance=self)
         return result
 
-    async def load(self) -> "Model":
+    async def load(self: T) -> T:
         """
         Allow to refresh existing Models fields from database.
         Be careful as the related models can be overwritten by pk_only models in load.
@@ -272,8 +275,8 @@ class Model(ModelRow):
         return self
 
     async def load_all(
-        self, follow: bool = False, exclude: Union[List, str, Set, Dict] = None
-    ) -> "Model":
+        self: T, follow: bool = False, exclude: Union[List, str, Set, Dict] = None
+    ) -> T:
         """
         Allow to refresh existing Models fields from database.
         Performs refresh of the related models fields.
@@ -305,7 +308,6 @@ class Model(ModelRow):
         if follow:
             relations = self._iterate_related_models()
         queryset = self.__class__.objects
-        print(relations)
         if exclude:
             queryset = queryset.exclude_fields(exclude)
         instance = await queryset.select_related(relations).get(pk=self.pk)